+++ lib/src/features/friends/data/repository/friends_store_repository_impl.dart
@@ -1,3 +1,17 @@
+import 'package:chat_application/src/core/constants/cloud_firestore_path.dart';
 import 'package:chat_application/src/features/friends/domain/repository/friends_store_repository.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
-class FriendsStoreRepositoryImpl extends FriendsStoreRepository {}
+class FriendsStoreRepositoryImpl extends FriendsStoreRepository {
+  @override
+  Stream<List> getFriendssSnapshot() {
+    User user = FirebaseAuth.instance.currentUser!;
+
+    return FirebaseFirestore.instance
+        .collection(CloudFirestorePath.friends)
+        .doc(user.uid)
+        .snapshots()
+        .map((event) => event.data()!['uids']);
+  }
+}
