+++ lib/src/features/friends/domain/repository/friends_store_repository.dart
@@ -1,1 +1,3 @@
-abstract class FriendsStoreRepository {}
+abstract class FriendsStoreRepository {
+  Stream<List> getFriendssSnapshot();
+}
