+++ lib/src/features/friends/domain/repository/users_store_repository.dart
@@ -1,7 +1,6 @@
 import 'package:chat_application/src/features/friends/domain/model/user_info.dart';
 
 abstract class UsersStoreRepository {
-  Stream<List> getFriendssSnapshot();
   Stream<UserInfo> getMyUserInfoSnapshot();
   Stream<UserInfo> getUserInfoSnapshot(String uid);
 }
